requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.1.2
django-anymail==12.0
django-ckeditor==6.7.0
django-ckeditor-5==0.2.10
django-cors-headers==3.14.0
django-dotenv==1.4.2
django-environ==0.9.0
django-import-export==3.2.0
django-jazzmin==2.6.0
django-js-asset==2.2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
et-xmlfile==1.1.0
gunicorn==23.0.0
humanize==4.6.0
idna==3.10
inflection==0.5.1
odfpy==1.4.1
openpyxl==3.1.5
packaging==24.1
pillow==11.0.0
PyJWT==2.9.0
pytz==2024.2
PyYAML==6.0.2
requests==2.31.0
shortuuid==1.0.11
sqlparse==0.4.4
tablib==3.7.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==6.7.0
xlrd==2.0.1
xlwt==1.3.0
